Updates tests to work with latest refactoring
